chore: Modified index page
